remove int casting on x positions communicated over the network

+++ src/main/java/com/mycodefu/werekitten/pipeline/handlers/network/ClientHandler.java
@@ -39,7 +39,7 @@ public ClientHandler() {
 				break;
 			}
 			case readyForInitMessage: {
-				nettyClient.sendMessage("init"+(int)context.level().get().getPixelScaleHelper().scaleXBack(context.level().get().getPlayerElement().getLocation().getX()));
+				nettyClient.sendMessage("init"+context.level().get().getPixelScaleHelper().scaleXBack(context.level().get().getPlayerElement().getLocation().getX()));
 				break;
 			}
